fix(Interaction): use local-space attach point in RotateTransformGrabAttach
This enables the ArtificialRotator to be used in moving objects (e.g. as a
steering wheel in a car), by making sure that movement of the
ArtificialRotator's parent object isn't factored into the rotation angle
calculation, which caused it to register as controller movement even
though the controller is kept still.